Creates Plottable interface for everything can be put in a plot
This introduces `Plottable` interface, to be implemented by everything
can be translated into a valid Gnuplot command (i.e., can be represented
		in plot). E.g., data files, labels.

Having created Plottable, I renamed PlottableItem, which is an abstract
superclass for DataSet and Function, into PlottableData. Indeed it
implements Plottable, but differs from a generic Plottable because it
holds data.